Initially upload
I had a standard version of this file that worked from the IDE and when running from the terminal, but when opening from finder (on macOS) it would save a new file and append to that file, as well as the reverse file, in the user directory, which was not what the program was expected to do.  I made some changes to get it to work with files from the same directory as the compiled file.